portal-elixir v0.0.1 Portal.Door

Functions

get(door)

Get the data currently in the door.

pop(door)

Pops a value from the door.

Returns {:ok, value} if there is a value or :error if the hole is currently empty.

push(door, value)

Pushes value into the door.

start_link(color)

Starts a door with the given color.

The color is given as a name so we can identify the door by color name instead of using a PID.